Decimal -> double

+++ PostgreEx.Core/Settings/CpuTupleCost.cs
@@ -1,6 +1,6 @@
 ﻿namespace PostgreEx.Core.Settings;
 
-public record CpuTupleCost(decimal Value)
+public record CpuTupleCost(double Value)
 {
-    public static implicit operator decimal(CpuTupleCost cpuTupleCost) => cpuTupleCost.Value;
+    public static implicit operator double(CpuTupleCost cpuTupleCost) => cpuTupleCost.Value;
 }
